server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["adi790uu <adi37068@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pydantic = "^2.10.2"
pydantic-settings = "^2.6.1"
loguru = "^0.7.2"
pymongo = ">4.9,<4.10"
motor = "^3.6.0"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
